+++ src/walksync.js
@@ -25,41 +25,43 @@
  * @api:Param   {Function}  called on every directory
  *                          function( fullPath, commonPath, fs.stat )
  *
+ * @api:Param   {Number}    DepthLimit - Maximum depth of children directories to search
+ *                                      -1 (Unlimited)
+ *                                       0 (no children)
+ *
  * @api:Example:
  *      walksync('/home/archivistnerd/')
  */
-module.exports = function walksync( base='', fileFn, pathFn ){
-  if ( !base.length ) throw new Error('blank path')
+module.exports = function walksync( base='', fileFn, pathFn, depthLimit=-1 ){
+  if ( !base.length )         throw new Error('blank path')
   if ( !fs.existsSync(base) ) throw new Error(`Path "${base}" does not exist`)
 
-  base = base.replace('\\','/')
-  if (base.substr(-1)!=='/') base+= '/'
-
-  let files = []
+  let files   = []
+    , folders = [ 
+                  ['', '', depthLimit]
+                ]
     ;
+  base = base.replace('\\','/') + ( (base.substr(-1)!=='/')? '/':'' )
 
-  function walk( path='', fullPath=(base+path), folders=[] ){
-    fullPath = fullPath.replace('\\','/')
-    if (fullPath.substr(-1)!=='/') fullPath+= '/'
+  while (folders.length){
+    let details  = folders.shift()
+      , path     = details[1]
+      , depth    = details.pop()
+      , fullPath = (base+path).replace('\\','/')
+      ;
+    if ( path !== '' && isFunction(pathFn) ) pathFn( ...details )
 
     fs.readdirSync( fullPath )
       .forEach( file => {
         let stats = fs.statSync( fullPath+file )
           ;
         // check if file is a directory
-        if ( stats.isDirectory() ) return folders.push( [fullPath+file, path+file, stats] )
-        // is a file check for callback function
+        if ( stats.isDirectory() ) return (depth!==0)? folders.push( [fullPath+file, path+file+'/', stats, depthLimit-1 ] ):0
+        // check for callback function
         if ( isFunction(fileFn) ) fileFn(fullPath+file, path+file, stats)
         files.push( path+file )
       });
-
-    folders.forEach( details => {
-        if ( isFunction(pathFn) ) pathFn( ...details )
-        walk( details[1]+'/' )
-      })
   }
-
-  walk()
 
   return files
 }