+++ docs/HISTORY.md
@@ -6,4 +6,6 @@ Archivist Nerd's walksync
 ## History
 |  Date     | Version  |    type   | Description     |
 |-----------|----------|-----------|-----------------|
+| 201120.14 |**v0.0.1-alpha-201120.1200**| **added** | depthLimit |
+| 201120.12 |**v0.0.1-alpha-201120.1200**| **refactored** | Refactored walksync function to remove recursive functions. |
 | 201119.23 |**v0.0.1-alpha**| **start** | project started |